refactor(listen): split upcoming show card helpers
